+++ NEWS.md
@@ -1,5 +1,10 @@
 # eummd
 
+
+## eummd 0.2.0, October 2024
+
+  - Fixed bug with meammd, and computation of p-value.
+
 ## eummd 0.1.8, May 2024
 
   - Added parameters `allowzeropval` and `alternative`